Move MixedRealityInputModule inspector to an editor project
