+++ Assets/MixedRealityToolkit.Services/InputSystem/Inspectors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9877a3079e968a4e97a2b371a0df9de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
