+++ Assets/MixedRealityToolkit.Services/InputSystem/Inspectors/MixedRealityInputModuleInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 732a086bd5e619443baf1720542c1b70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
