+++ Assets/MixedRealityToolkit.Services/InputSystem/Inspectors/MixedRealityToolkit.Services.Inspectors.asmdef
@@ -1,0 +1,13 @@
+{
+    "name": "Microsoft.MixedReality.Toolkit.Services.InputSystem.Inspectors",
+    "references": [
+        "Microsoft.MixedReality.Toolkit",
+        "Microsoft.MixedReality.Toolkit.Services.InputSystem"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false
+}
