+++ Assets/MixedRealityToolkit.Services/InputSystem/Inspectors/MixedRealityToolkit.Services.Inspectors.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8eb5c41a66d79584997ac78c84bd18fb
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
